# test_openapi.v2.0.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Flask Requests",
    "description": "TODO: Add Description"
  },
  "host": "127.0.0.1:7777",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/users/1": {
      "get": {
        "description": "TODO: Add Description",
        "summary": "Get User 1",
        "tags": [
          "Misc"
        ],
        "operationId": "Users1Get",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Accept",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-unitTests": [
          {
            "request": {
              "method": "GET",
              "uri": "/users/1",
              "headers": {
                "Accept": "application/json"
              }
            },
            "expectedResponse": {
              "x-allowExtraHeaders": true,
              "x-bodyMatchMode": "NONE",
              "x-arrayOrderedMatching": false,
              "x-arrayCheckCount": false,
              "x-matchResponseSchema": true,
              "headers": {}
            },
            "x-testShouldPass": true,
            "x-testEnabled": true,
            "x-testName": "Get User 1",
            "x-testDescription": "TODO: Add Description"
          }
        ],
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        }
      },
      "put": {
        "description": "TODO: Add Description",
        "summary": "Update User",
        "tags": [
          "Misc"
        ],
        "operationId": "Users1Put",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Accept",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateUserrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-unitTests": [
          {
            "request": {
              "method": "PUT",
              "uri": "/users/1",
              "headers": {
                "Accept": "application/json",
                "Content-Type": "application/json"
              },
              "body": "{\"user\":\"admin\",\"first\":\"john\",\"last\":\"smith\",\"password\":\"hello\"}"
            },
            "expectedResponse": {
              "x-allowExtraHeaders": true,
              "x-bodyMatchMode": "NONE",
              "x-arrayOrderedMatching": false,
              "x-arrayCheckCount": false,
              "x-matchResponseSchema": true,
              "headers": {}
            },
            "x-testShouldPass": true,
            "x-testEnabled": true,
            "x-testName": "Update User",
            "x-testDescription": "TODO: Add Description"
          }
        ],
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        }
      }
    },
    "/users": {
      "get": {
        "description": "TODO: Add Description",
        "summary": "Get Users by Name",
        "tags": [
          "Misc"
        ],
        "operationId": "UsersGet2",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "user",
            "in": "query",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Accept",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-unitTests": [
          {
            "request": {
              "method": "GET",
              "uri": "/users?user=dd",
              "headers": {
                "Accept": "application/json"
              }
            },
            "expectedResponse": {
              "x-allowExtraHeaders": true,
              "x-bodyMatchMode": "NONE",
              "x-arrayOrderedMatching": false,
              "x-arrayCheckCount": false,
              "x-matchResponseSchema": true,
              "headers": {}
            },
            "x-testShouldPass": true,
            "x-testEnabled": true,
            "x-testName": "Get Users by Name",
            "x-testDescription": "TODO: Add Description"
          }
        ],
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        }
      },
      "post": {
        "description": "TODO: Add Description",
        "summary": "Create User",
        "tags": [
          "Misc"
        ],
        "operationId": "UsersPost",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Accept",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/CreateUserrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-unitTests": [
          {
            "request": {
              "method": "POST",
              "uri": "/users",
              "headers": {
                "Accept": "application/json",
                "Content-Type": "application/json"
              },
              "body": "{\"user\":\"dd\",\"first\":\"mike\",\"last\":\"smith\",\"password\":\"hello\"}"
            },
            "expectedResponse": {
              "x-allowExtraHeaders": true,
              "x-bodyMatchMode": "NONE",
              "x-arrayOrderedMatching": false,
              "x-arrayCheckCount": false,
              "x-matchResponseSchema": true,
              "headers": {}
            },
            "x-testShouldPass": true,
            "x-testEnabled": true,
            "x-testName": "Create User",
            "x-testDescription": "TODO: Add Description"
          }
        ],
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        }
      }
    },
    "/users/2": {
      "delete": {
        "description": "TODO: Add Description",
        "summary": "Delete User 2",
        "tags": [
          "Misc"
        ],
        "operationId": "Users2Delete",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Accept",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-unitTests": [
          {
            "request": {
              "method": "DELETE",
              "uri": "/users/2",
              "headers": {
                "Accept": "application/json"
              }
            },
            "expectedResponse": {
              "x-allowExtraHeaders": true,
              "x-bodyMatchMode": "NONE",
              "x-arrayOrderedMatching": false,
              "x-arrayCheckCount": false,
              "x-matchResponseSchema": true,
              "headers": {}
            },
            "x-testShouldPass": true,
            "x-testEnabled": true,
            "x-testName": "Delete User 2",
            "x-testDescription": "TODO: Add Description"
          }
        ],
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        }
      }
    }
  },
  "definitions": {
    "CreateUserrequest": {
      "title": "Create UserRequest",
      "example": {
        "user": "dd",
        "first": "mike",
        "last": "smith",
        "password": "hello"
      },
      "type": "object",
      "properties": {
        "user": {
          "description": "",
          "example": "dd",
          "type": "string"
        },
        "first": {
          "description": "",
          "example": "mike",
          "type": "string"
        },
        "last": {
          "description": "",
          "example": "smith",
          "type": "string"
        },
        "password": {
          "description": "",
          "example": "hello",
          "type": "string"
        }
      },
      "required": [
        "user",
        "first",
        "last",
        "password"
      ]
    },
    "UpdateUserrequest": {
      "title": "Update UserRequest",
      "example": {
        "user": "admin",
        "first": "john",
        "last": "smith",
        "password": "hello"
      },
      "type": "object",
      "properties": {
        "user": {
          "description": "",
          "example": "admin",
          "type": "string"
        },
        "first": {
          "description": "",
          "example": "john",
          "type": "string"
        },
        "last": {
          "description": "",
          "example": "smith",
          "type": "string"
        },
        "password": {
          "description": "",
          "example": "hello",
          "type": "string"
        }
      },
      "required": [
        "user",
        "first",
        "last",
        "password"
      ]
    }
  }
}